Complete q_size function
the q_size function returns size of queue which means the total nodes number in queue.
The function implemented by increase length counter variable which is len in list_for_each loop. If queue is empty, then the function returns 0 while checking whether head pointer is NULL in the beginning.